Test Update 1.0

+++ Calculation.cpp
@@ -5,9 +5,9 @@
 
 struct Solution SolveQuadraticEquation(struct Equation * coefficient)
 {
-    float coefficient_a = coefficient->a, coefficient_b = coefficient->b, coefficient_c = coefficient->c;
-    float discriminant =  coefficient_b * coefficient_b  - 4 * coefficient_a * coefficient_c;
-    float sqrt_D = sqrt(discriminant);
+    double coefficient_a = coefficient->a, coefficient_b = coefficient->b, coefficient_c = coefficient->c;
+    double discriminant =  coefficient_b * coefficient_b  - 4 * coefficient_a * coefficient_c;
+    double sqrt_D = sqrt(discriminant);
     struct Solution out = {.solution_type = EQUATION_TYPE_NOSOLUTION, .solution_1 = 0, .solution_2 = 0};
 
     if (!IsNull(coefficient_a))
